refactor(approval-gate): extract intercept.rs from lib.rs
Fourth step of breaking lib.rs into focused modules. This commit
lifts the intercept decision flow into its own module — the three
layers that decide what the gate does with an incoming function
call: policy rules, per-function interceptor rules, and classifier
replies, plus the async handle_intercept that writes the pending
record.

Moved (lib.rs -> intercept.rs):
- InterceptAction enum + decide_intercept_action()
- PolicyOutcome enum + apply_policy_rules()
- ClassifierDecision enum + interpret_classifier_reply()
- handle_intercept()

handle_intercept stays public (re-exported from the crate root).
The pub(crate) helpers stay pub(crate) and are imported back into
lib.rs for the subscriber closure in register().

Pure move, no behavior change. 141 approval-gate tests pass;
turn-orchestrator + harness + shell verified green.
